Remove Keycloak auth and replace with LocalAuthService (closes #134)
