+++ src/NoteBookmark.MauiApp/Platforms/Android/AndroidManifest.xml
@@ -1,15 +1,6 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application android:allowBackup="true" android:icon="@mipmap/appicon" android:roundIcon="@mipmap/appicon_round" android:supportsRtl="true">
-        <!-- Intent filter required by WebAuthenticator to handle Keycloak OIDC redirect -->
-        <activity android:name="microsoft.maui.essentials.WebAuthenticatorCallbackActivity" android:exported="true">
-            <intent-filter android:label="notebookmark auth callback">
-                <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="notebookmark" android:host="auth" android:pathPrefix="/callback" />
-            </intent-filter>
-        </activity>
     </application>
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
