+++ src/NoteBookmark.MauiApp/wwwroot/appsettings.json
@@ -1,7 +1,2 @@
 {
-  "Keycloak": {
-    "Authority": "https://{YOUR_KEYCLOAK_HOST}/realms/notebookmark",
-    "ClientId": "notebookmark-mobile",
-    "RedirectUri": "notebookmark://auth/callback"
-  }
 }
